添加更新进度时修改缓存的收视列表中对应条目的LastToush，将所有的LastToush改为long

+++ Bangumi.Api/Models/Watching.cs
@@ -21,7 +21,7 @@ namespace Bangumi.Api.Models
         public int VolStatus { get; set; }
 
         [JsonProperty("lasttouch")]
-        public int LastTouch { get; set; }
+        public long LastTouch { get; set; }
 
         [JsonProperty("subject")]
         public Subject3 Subject { get; set; }
